Refactor price info tests
