+++ data/products.ts
@@ -1,4 +1,4 @@
-type Product = {
+export type Product = {
   id: number;
   title: string;
   imgSrc: string;
