+++ tests/checkoutOverviewPage.spec.ts
@@ -219,25 +219,11 @@ test.describe('Items purchased', () => {
         });
 
         test('Price total info reflects cost of items purchased', async ({ page }) => {
-          // Sort products into random order using the Fisher-Yates algorithm
-          let products = [...PRODUCT_INFO];
-          let j = products.length;
-          while (j) {
-            const i = Math.floor(Math.random() * j--);
-            [products[j], products[i]] = [products[i], products[j]];
-          }
-          // Slice off first n elements of the array
-          const purchasedProducts = products.slice(0, Math.ceil(Math.random() * products.length));
-          await setCartContentsInLocalStorage(
-            page,
-            purchasedProducts.map((prod) => prod.id),
-            URLS.checkoutOverviewPage
-          );
+          const purchasedProducts = await checkoutOverviewPage.addRandomProductsToCart(page);
 
           const subtotalAmount = purchasedProducts.reduce((n, { price }) => n + Number(price), 0);
-          // Tax is charged at 8%
-          const taxAmount = (Math.round(subtotalAmount * 100 * 0.08) / 100).toFixed(2);
-          const totalAmount = (Math.round((Number(subtotalAmount) + Number(taxAmount)) * 100) / 100).toFixed(2);
+          const taxAmount = checkoutOverviewPage.calculateTax(subtotalAmount);
+          const totalAmount = checkoutOverviewPage.calculateTotal(subtotalAmount, Number(taxAmount));
           await expect(checkoutOverviewPage.subtotalLabel).toHaveText(
             `${EXPECTED_TEXT.priceInfo.subtotalPrefix}${subtotalAmount}`
           );
@@ -281,27 +267,14 @@ test.describe('Items purchased', () => {
 
         test('Price total info does not accurately reflect cost of items purchased', async ({ page }) => {
           const subtotalRegex = '\\d{1,3}(.\\d{1,2})?$';
-          // Sort products into random order using the Fisher-Yates algorithm
-          let products = [...PRODUCT_INFO];
-          let j = products.length;
-          while (j) {
-            const i = Math.floor(Math.random() * j--);
-            [products[j], products[i]] = [products[i], products[j]];
-          }
-          // Slice off first n elements of the array
-          const purchasedProducts = products.slice(0, Math.ceil(Math.random() * products.length));
-          await setCartContentsInLocalStorage(
-            page,
-            purchasedProducts.map((prod) => prod.id),
-            URLS.checkoutOverviewPage
-          );
+          // Although this method returns an array of products we don't use that array in the rest of the test
+          await checkoutOverviewPage.addRandomProductsToCart(page);
 
           const subtotalStr = await checkoutOverviewPage.subtotalLabel.textContent();
           expect(subtotalStr).toMatch(new RegExp(`${subtotalRegex}`));
           const subtotalAmount = Number(subtotalStr!.replace(EXPECTED_TEXT.priceInfo.subtotalPrefix, ''));
-          // Tax is charged at 8%
-          const taxAmount = (Math.round(subtotalAmount * 100 * 0.08) / 100).toFixed(2);
-          const totalAmount = (Math.round((Number(subtotalAmount) + Number(taxAmount)) * 100) / 100).toFixed(2);
+          const taxAmount = checkoutOverviewPage.calculateTax(subtotalAmount);
+          const totalAmount = checkoutOverviewPage.calculateTotal(subtotalAmount, Number(taxAmount));
           await expect(checkoutOverviewPage.taxLabel).toHaveText(`${EXPECTED_TEXT.priceInfo.taxPrefix}${taxAmount}`);
           await expect(checkoutOverviewPage.totalLabel).toHaveText(
             `${EXPECTED_TEXT.priceInfo.totalPrefix}${totalAmount}`
